Moved plot tests into subpackages
