+++ tests/plotting/bokeh/testannotationplot.py
@@ -1,0 +1,34 @@
+from holoviews.element import HLine, VLine
+
+from .testplot import TestBokehPlot, bokeh_renderer
+
+
+class TestHVLinePlot(TestBokehPlot):
+
+    def test_hline_invert_axes(self):
+        hline = HLine(1.1).opts(plot=dict(invert_axes=True))
+        plot = bokeh_renderer.get_plot(hline)
+        span = plot.handles['glyph']
+        self.assertEqual(span.dimension, 'height')
+        self.assertEqual(span.location, 1.1)
+
+    def test_hline_plot(self):
+        hline = HLine(1.1)
+        plot = bokeh_renderer.get_plot(hline)
+        span = plot.handles['glyph']
+        self.assertEqual(span.dimension, 'width')
+        self.assertEqual(span.location, 1.1)
+
+    def test_vline_invert_axes(self):
+        vline = VLine(1.1).opts(plot=dict(invert_axes=True))
+        plot = bokeh_renderer.get_plot(vline)
+        span = plot.handles['glyph']
+        self.assertEqual(span.dimension, 'width')
+        self.assertEqual(span.location, 1.1)
+
+    def test_vline_plot(self):
+        vline = VLine(1.1)
+        plot = bokeh_renderer.get_plot(vline)
+        span = plot.handles['glyph']
+        self.assertEqual(span.dimension, 'height')
+        self.assertEqual(span.location, 1.1)
