+++ tests/plotting/bokeh/testelementplot.py
@@ -1,0 +1,79 @@
+from nose.plugins.attrib import attr
+
+import numpy as np
+
+from holoviews.core import Dimension, DynamicMap
+from holoviews.element import Curve, Image
+from holoviews.streams import Stream
+
+from .testplot import TestBokehPlot, bokeh_renderer
+
+try:
+    from bokeh.document import Document
+    from bokeh.models import FuncTickFormatter
+except:
+    pass
+
+
+
+class TestElementPlotPlot(TestBokehPlot):
+
+    def test_element_show_frame_disabled(self):
+        curve = Curve(range(10)).opts(plot=dict(show_frame=False))
+        plot = bokeh_renderer.get_plot(curve).state
+        self.assertEqual(plot.outline_line_alpha, 0)
+
+    def test_element_no_xaxis(self):
+        curve = Curve(range(10)).opts(plot=dict(xaxis=None))
+        plot = bokeh_renderer.get_plot(curve).state
+        self.assertFalse(plot.xaxis[0].visible)
+
+    def test_element_no_yaxis(self):
+        curve = Curve(range(10)).opts(plot=dict(yaxis=None))
+        plot = bokeh_renderer.get_plot(curve).state
+        self.assertFalse(plot.yaxis[0].visible)
+
+    def test_element_xrotation(self):
+        curve = Curve(range(10)).opts(plot=dict(xrotation=90))
+        plot = bokeh_renderer.get_plot(curve).state
+        self.assertEqual(plot.xaxis[0].major_label_orientation, np.pi/2)
+
+    def test_element_yrotation(self):
+        curve = Curve(range(10)).opts(plot=dict(yrotation=90))
+        plot = bokeh_renderer.get_plot(curve).state
+        self.assertEqual(plot.yaxis[0].major_label_orientation, np.pi/2)
+
+    def test_static_source_optimization(self):
+        global data
+        data = np.ones((5, 5))
+        img = Image(data)
+        def get_img(test):
+            global data
+            data *= test
+            return img
+        stream = Stream.define(str('Test'), test=1)()
+        dmap = DynamicMap(get_img, streams=[stream])
+        plot = bokeh_renderer.get_plot(dmap, doc=Document())
+        source = plot.handles['source']
+        self.assertEqual(source.data['image'][0].mean(), 1)
+        stream.event(test=2)
+        self.assertTrue(plot.static_source)
+        self.assertEqual(source.data['image'][0].mean(), 2)
+        self.assertNotIn(source, plot.current_handles)
+
+    @attr(optional=1)  # Requires Flexx
+    def test_element_formatter_xaxis(self):
+        def formatter(x):
+            return '%s' % x
+        curve = Curve(range(10), kdims=[Dimension('x', value_format=formatter)])
+        plot = bokeh_renderer.get_plot(curve).state
+        self.assertIsInstance(plot.xaxis[0].formatter, FuncTickFormatter)
+
+    @attr(optional=1)  # Requires Flexx
+    def test_element_formatter_yaxis(self):
+        def formatter(x):
+            return '%s' % x
+        curve = Curve(range(10), vdims=[Dimension('y', value_format=formatter)])
+        plot = bokeh_renderer.get_plot(curve).state
+        self.assertIsInstance(plot.yaxis[0].formatter, FuncTickFormatter)
+
