+++ tests/plotting/matplotlib/testpointplot.py
@@ -1,0 +1,59 @@
+import numpy as np
+
+from holoviews.element import Points
+
+from .testplot import TestMPLPlot, mpl_renderer
+from ..utils import ParamLogStream
+
+try:
+    from matplotlib import pyplot
+except:
+    pass
+
+
+class TestPointPlot(TestMPLPlot):
+
+    def test_points_non_numeric_size_warning(self):
+        data = (np.arange(10), np.arange(10), list(map(chr, range(94,104))))
+        points = Points(data, vdims=['z']).opts(plot=dict(size_index=2))
+        with ParamLogStream() as log:
+            plot = mpl_renderer.get_plot(points)
+        log_msg = log.stream.read()
+        warning = ('%s: z dimension is not numeric, '
+                   'cannot use to scale Points size.\n' % plot.name)
+        self.assertEqual(log_msg, warning)
+
+    def test_points_cbar_extend_both(self):
+        img = Points(([0, 1], [0, 3])).redim(y=dict(range=(1,2)))
+        plot = mpl_renderer.get_plot(img(plot=dict(colorbar=True, color_index=1)))
+        self.assertEqual(plot.handles['cbar'].extend, 'both')
+
+    def test_points_cbar_extend_min(self):
+        img = Points(([0, 1], [0, 3])).redim(y=dict(range=(1, None)))
+        plot = mpl_renderer.get_plot(img(plot=dict(colorbar=True, color_index=1)))
+        self.assertEqual(plot.handles['cbar'].extend, 'min')
+
+    def test_points_cbar_extend_max(self):
+        img = Points(([0, 1], [0, 3])).redim(y=dict(range=(None, 2)))
+        plot = mpl_renderer.get_plot(img(plot=dict(colorbar=True, color_index=1)))
+        self.assertEqual(plot.handles['cbar'].extend, 'max')
+
+    def test_points_cbar_extend_clime(self):
+        img = Points(([0, 1], [0, 3])).opts(style=dict(clim=(None, None)))
+        plot = mpl_renderer.get_plot(img(plot=dict(colorbar=True, color_index=1)))
+        self.assertEqual(plot.handles['cbar'].extend, 'neither')
+
+    def test_points_rcparams_do_not_persist(self):
+        opts = dict(fig_rcparams={'text.usetex': True})
+        points = Points(([0, 1], [0, 3])).opts(plot=opts)
+        mpl_renderer.get_plot(points)
+        self.assertFalse(pyplot.rcParams['text.usetex'])
+
+    def test_points_rcparams_used(self):
+        opts = dict(fig_rcparams={'grid.color': 'red'})
+        points = Points(([0, 1], [0, 3])).opts(plot=opts)
+        plot = mpl_renderer.get_plot(points)
+        ax = plot.state.axes[0]
+        lines = ax.get_xgridlines()
+        self.assertEqual(lines[0].get_color(), 'red')
+
