+++ tests/plotting/plotly/testplot.py
@@ -1,0 +1,114 @@
+from collections import deque
+from unittest import SkipTest
+from nose.plugins.attrib import attr
+
+import numpy as np
+
+from holoviews.core import Store, DynamicMap, GridSpace
+from holoviews.element import Curve, Scatter3D, Image
+from holoviews.element.comparison import ComparisonTestCase
+from holoviews.streams import PointerX
+from holoviews.plotting import comms
+
+try:
+    import holoviews.plotting.plotly # noqa (Activate backend)
+    plotly_renderer = Store.renderers['plotly']
+except:
+    plotly_renderer = None
+
+
+@attr(optional=1)
+class TestPlotlyPlotInstantiation(ComparisonTestCase):
+
+    def setUp(self):
+        self.previous_backend = Store.current_backend
+        Store.current_backend = 'plotly'
+        self.default_comm = plotly_renderer.comms['default']
+        if not plotly_renderer:
+            raise SkipTest("Plotly required to test plot instantiation")
+        plotly_renderer.comms['default'] = (comms.Comm, '')
+
+    def tearDown(self):
+        Store.current_backend = self.previous_backend
+        plotly_renderer.comms['default'] = self.default_comm
+
+    def _get_plot_state(self, element):
+        plot = plotly_renderer.get_plot(element)
+        plot.initialize_plot()
+        return plot.state
+
+    def test_curve_state(self):
+        curve = Curve([1, 2, 3])
+        state = self._get_plot_state(curve)
+        self.assertEqual(state['data'][0]['y'], np.array([1, 2, 3]))
+        self.assertEqual(state['layout']['yaxis']['range'], [1, 3])
+
+    def test_scatter3d_state(self):
+        scatter = Scatter3D(([0,1], [2,3], [4,5]))
+        state = self._get_plot_state(scatter)
+        self.assertEqual(state['data'][0]['x'], np.array([0, 1]))
+        self.assertEqual(state['data'][0]['y'], np.array([2, 3]))
+        self.assertEqual(state['data'][0]['z'], np.array([4, 5]))
+        self.assertEqual(state['layout']['scene']['xaxis']['range'], [0, 1])
+        self.assertEqual(state['layout']['scene']['yaxis']['range'], [2, 3])
+        self.assertEqual(state['layout']['scene']['zaxis']['range'], [4, 5])
+
+    def test_overlay_state(self):
+        layout = Curve([1, 2, 3]) * Curve([2, 4, 6])
+        state = self._get_plot_state(layout)
+        self.assertEqual(state['data'][0]['y'], np.array([1, 2, 3]))
+        self.assertEqual(state['data'][1]['y'], np.array([2, 4, 6]))
+        self.assertEqual(state['layout']['yaxis']['range'], [1, 6])
+
+    def test_layout_state(self):
+        layout = Curve([1, 2, 3]) + Curve([2, 4, 6])
+        state = self._get_plot_state(layout)
+        self.assertEqual(state['data'][0]['y'], np.array([1, 2, 3]))
+        self.assertEqual(state['data'][0]['yaxis'], 'y1')
+        self.assertEqual(state['data'][1]['y'], np.array([2, 4, 6]))
+        self.assertEqual(state['data'][1]['yaxis'], 'y2')
+
+    def test_grid_state(self):
+        grid = GridSpace({(i, j): Curve([i, j]) for i in [0, 1]
+                          for j in [0, 1]})
+        state = self._get_plot_state(grid)
+        self.assertEqual(state['data'][0]['y'], np.array([0, 0]))
+        self.assertEqual(state['data'][0]['xaxis'], 'x1')
+        self.assertEqual(state['data'][0]['yaxis'], 'y1')
+        self.assertEqual(state['data'][1]['y'], np.array([1, 0]))
+        self.assertEqual(state['data'][1]['xaxis'], 'x2')
+        self.assertEqual(state['data'][1]['yaxis'], 'y1')
+        self.assertEqual(state['data'][2]['y'], np.array([0, 1]))
+        self.assertEqual(state['data'][2]['xaxis'], 'x1')
+        self.assertEqual(state['data'][2]['yaxis'], 'y2')
+        self.assertEqual(state['data'][3]['y'], np.array([1, 1]))
+        self.assertEqual(state['data'][3]['xaxis'], 'x2')
+        self.assertEqual(state['data'][3]['yaxis'], 'y2')
+
+    def test_stream_callback_single_call(self):
+        def history_callback(x, history=deque(maxlen=10)):
+            history.append(x)
+            return Curve(list(history))
+        stream = PointerX(x=0)
+        dmap = DynamicMap(history_callback, kdims=[], streams=[stream])
+        plot = plotly_renderer.get_plot(dmap)
+        plotly_renderer(plot)
+        for i in range(20):
+            stream.event(x=i)
+        state = plot.state
+        self.assertEqual(state['data'][0]['x'], np.arange(10))
+        self.assertEqual(state['data'][0]['y'], np.arange(10, 20))
+
+    def test_layout_instantiate_subplots(self):
+        layout = (Curve(range(10)) + Curve(range(10)) + Image(np.random.rand(10,10)) +
+                  Curve(range(10)) + Curve(range(10)))
+        plot = plotly_renderer.get_plot(layout)
+        positions = [(0, 0), (0, 1), (0, 2), (0, 3), (1, 0), (1, 1), (1, 2), (1, 3)]
+        self.assertEqual(sorted(plot.subplots.keys()), positions)
+
+    def test_layout_instantiate_subplots_transposed(self):
+        layout = (Curve(range(10)) + Curve(range(10)) + Image(np.random.rand(10,10)) +
+                  Curve(range(10)) + Curve(range(10)))
+        plot = plotly_renderer.get_plot(layout(plot=dict(transpose=True)))
+        positions = [(0, 0), (0, 1), (1, 0), (1, 1), (2, 0), (2, 1), (3, 0), (3, 1)]
+        self.assertEqual(sorted(plot.subplots.keys()), positions)
